+++ core-ng-test/src/main/java/core/framework/module/TestDBConfig.java
@@ -1,6 +1,5 @@
 package core.framework.module;
 
-import core.framework.impl.db.Vendor;
 import core.framework.util.Strings;
 
 /**
@@ -8,20 +7,8 @@ import core.framework.util.Strings;
  */
 public class TestDBConfig extends DBConfig {
     @Override
-    String databaseURL(String url, Vendor vendor) {
-        String syntaxParam = hsqldbSyntaxParam(vendor);
-        return Strings.format("jdbc:hsqldb:mem:{};{}", name == null ? "." : name, syntaxParam);
-    }
-
-    private String hsqldbSyntaxParam(Vendor vendor) {
-        switch (vendor) {
-            case ORACLE:
-                return "sql.syntax_ora=true";
-            case MYSQL:
-                return "sql.syntax_mys=true";
-            default:
-                return "";
-        }
+    String databaseURL(String url) {
+        return Strings.format("jdbc:hsqldb:mem:{};sql.syntax_mys=true", name == null ? "." : name);
     }
 
     @Override