+++ core-ng/src/main/java/core/framework/db/PrimaryKey.java
@@ -14,6 +14,4 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 public @interface PrimaryKey {
     // only allow auto increment on primary key, not regular column
     boolean autoIncrement() default false;
-
-    String sequence() default "";
 }