+++ core-ng/src/main/java/core/framework/impl/db/Vendor.java
@@ -1,8 +1,0 @@
-package core.framework.impl.db;
-
-/**
- * @author neo
- */
-public enum Vendor {
-    MYSQL, ORACLE
-}