+++ core-ng/src/main/java/core/framework/internal/db/UpdateQuery.java
@@ -1,0 +1,18 @@
+package core.framework.internal.db;
+
+/**
+ * @author neo
+ */
+public interface UpdateQuery<T> {
+    Statement update(T entity, boolean partial);
+
+    class Statement {
+        final String sql;
+        final Object[] params;
+
+        Statement(String sql, Object... params) {
+            this.sql = sql;
+            this.params = params;
+        }
+    }
+}