+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 7.0.0 (8/9/2019 - ) !!! only support jdk 12, recommended to update to 6.13.9 and all runtime to adoptopenjdk-12 before update to 7.0.0 !!! 
 * project: updated source/target to java 12
+* db: drop oracle support, only support mysql to simplify  
 
 ### 6.13.9 (8/8/2019)  !!! this is the last version for JDK 11, update to this version then start update all JDK runtime to 12+ !!! 
 * config: update controller inspector to adapt to JDK 12+